move loader stack to query-loader
