+++ addon/database/internal-factory.js
@@ -74,8 +74,8 @@ export default Ember.Mixin.create({
     return this.get('store')._createInternalFilter(this, owner, opts);
   },
 
-  _createInternalLoader(owner, opts, type) {
-    return this.get('store')._createInternalLoader(this, owner, opts, type);
+  _createInternalQueryLoader(owner, opts, type) {
+    return this.get('store')._createInternalQueryLoader(this, owner, opts, type);
   }
 
 });
