+++ addon/document/query-loader.js
